+++ Sharpnado.Presentation.Forms.iOS/Renderers/HorizontalList/iOSHorizontalListViewRenderer.cs
@@ -138,7 +138,7 @@ namespace Sharpnado.Presentation.Forms.iOS.Renderers.HorizontalList
             bool heightChanged = false;
             if (Element.ColumnCount > 0)
             {
-                int newItemWidth = Element.ComputeItemWidth(width);
+                double newItemWidth = Element.ComputeItemWidth(width);
                 if (Element.ItemWidth != newItemWidth)
                 {
                     Element.ItemWidth = newItemWidth;
@@ -148,7 +148,7 @@ namespace Sharpnado.Presentation.Forms.iOS.Renderers.HorizontalList
 
             if (Element.IsLayoutLinear && Element.ItemHeight == 0)
             {
-                int newItemHeight = Element.ComputeItemHeight(height);
+                double newItemHeight = Element.ComputeItemHeight(height);
                 if (Element.ItemHeight != newItemHeight)
                 {
                     Element.ItemHeight = newItemHeight;
